Implement a FakeBackend that raises in get_tiles
Use it to check that Context.load passes the backend. Also added
an xfail test case for #936
